+++ CloudHarvestCLI/commands/harvest/__init__.py
@@ -1,0 +1,1 @@
+from CloudHarvestCLI.commands.report.commandset import ReportCommand