+++ CloudHarvestCLI/commands/services/__init__.py
@@ -1,1 +1,0 @@
-from CloudHarvestCLI.commands.services.commandset import ServicesCommand