+++ CloudHarvestCLI/commands/theme/__init__.py
@@ -1,0 +1,1 @@
+from CloudHarvestCLI.commands.theme.commandset import ThemeCommand