+++ docker/docker-entrypoint.sh
@@ -21,10 +21,6 @@ do
         rebuild=1
         shift # Remove --rebuild from processing
         ;;
-        --config)
-        config=1
-        shift # Remove --config from processing
-        ;;
         --version)
         version=1
         shift # Remove --version from processing
@@ -36,8 +32,7 @@ do
         echo "--tag tag: Allows you to specify the Docker image tag."
         echo
         echo "Configuration:"
-        echo "--config: Run the configuration script to create the harvest.json file."
-        echo "--rebuild: Deletes the entire contents of the './app' directory. Implies --config."
+        echo "--rebuild: Deletes the entire contents of the './app' directory."
         echo "--version: Prints the version, commit hash, and branch name then exits."
         help_requested=1
         ;;
@@ -69,32 +64,12 @@ fi
 
 # Check if the --rebuild flag was provided; wipe the app directory if it was
 if [ $rebuild -eq 1 ]; then
-    rm -rf /src/appsource
-    mkdir -p /src/app
+    rm -rf /src/app
 fi
 
 if [ ! -d "/src/app" ]; then
     echo "Creating app directory."
     mkdir -p "/src/app"
-fi
-
-# Check if the app/harvest.json file exists or --config is provided
-if [ ! -f "/src/app/harvest.json" ] || [ $config -eq 1 ]; then
-    # If the file does not exist or --config is provided, start config.py using docker run
-    source /venv/bin/activate && python /src/config.py
-
-    # Check the exit status of config.py
-    if [ $? -ne 0 ]; then
-        # If the exit status is not 0, abort the script
-        echo "config.py exited with an error. Aborting."
-        exit 1
-    fi
-
-    # If --config was provided, exit with a status code of 0
-    if [ $config -eq 1 ] || [ $rebuild -eq 1 ]; then
-        echo "You may now restart Harvest"
-        exit 0
-    fi
 fi
 
 echo "Harvest is starting!" \